:recycle: Remove camera images from scanner component and rename the files to used standard convention
