+++ src/components/shared/_Scanner/Scanner.styles.js
