+++ src/components/shared/_Scanner/_CameraOverlay.js
@@ -8,7 +8,7 @@ import { colors } from 'constants/styleGuide';
 import GallerySvg from 'assets/svgs/GallerySvg';
 import { IconButton } from '../toolBox/button';
 import { P } from '../toolBox/typography';
-import getStyles from './styles';
+import getStyles from './Scanner.styles';
 
 const CameraOverlay = ({ close, toggleGallery }) => {
   const { styles, theme } = useTheme({ styles: getStyles() });
